# Optimize Compiler Settings