--- a/modules/passkey/contracts/4337/SafeWebAuthnSharedSigner.sol
+++ b/modules/passkey/contracts/4337/SafeWebAuthnSharedSigner.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: LGPL-3.0-only
-pragma solidity >=0.8.0;
+pragma solidity 0.8.26;
 
 import {SignatureValidator} from "../base/SignatureValidator.sol";
 import {ISafe} from "../interfaces/ISafe.sol";